Added the standings component
Prior to this commit, the website did not have the standings component.

This change adds the standings component, with sortable columns for GP,
W, L, T, GF and GA, along with the necessary changes to the Team model,
request interceptor and sample data.

This change also combined all services into one BackendService.